Creating the flow of execution between services and DB calls and API call. Need to make create call to DB now.

+++ LancerWebAPI/Services/IWebsiteServices.cs
@@ -2,6 +2,6 @@
 {
     public interface IWebsiteServices
     {
-        public Task<IEnumerable<WebsiteModel>> GetAllPlaces(string location, string query, double distance);
+        public Task<IEnumerable<WebsiteModel>> GetAllPlaces(string location, string query, int distance);
     }
 }
